requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-filter==24.1
djangorestframework==3.14.0
drf-nested-routers==0.93.5
drf-yasg==1.21.7
importlib-metadata==7.0.0
inflection==0.5.1
itypes==1.2.0
Markdown==3.5.1
packaging==23.2
Pillow==10.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
zipp==3.17.0
